refactor(shared): ErrorResponse/ErrorCode を shared.web へ移動(設計規約 §2.4 / ADR-0011 対応)
shared.adapter.web から shared.web へパッケージ移動。TaskExceptionHandler のインポートを更新。
