+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/shared/web/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package xyz.dgz48.tasks.webapi.shared.web;
+
+import org.jspecify.annotations.NullMarked;
